+++ src/main/java/com/mcommings/campaigner/services/people/PersonService.java
@@ -4,9 +4,7 @@ import com.mcommings.campaigner.interfaces.people.IPerson;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.people.Person;
 import com.mcommings.campaigner.repositories.IWealthRepository;
-import com.mcommings.campaigner.repositories.people.IAbilityScoreRepository;
-import com.mcommings.campaigner.repositories.people.IPersonRepository;
-import com.mcommings.campaigner.repositories.people.IRaceRepository;
+import com.mcommings.campaigner.repositories.people.*;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -18,6 +16,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_PERSON;
 import static java.util.Objects.isNull;
 
 @Service
@@ -27,14 +26,19 @@ public class PersonService implements IPerson {
     private final IRaceRepository raceRepository;
     private final IWealthRepository wealthRepository;
     private final IAbilityScoreRepository abilityScoreRepository;
+    private final IJobAssignmentRepository jobAssignmentRepository;
+    private final IEventPlacePersonRepository eventPlacePersonRepository;
 
     @Autowired
     public PersonService(IPersonRepository personRepository, IRaceRepository raceRepository,
-                         IWealthRepository wealthRepository, IAbilityScoreRepository abilityScoreRepository) {
+                         IWealthRepository wealthRepository, IAbilityScoreRepository abilityScoreRepository,
+                         IJobAssignmentRepository jobAssignmentRepository, IEventPlacePersonRepository eventPlacePersonRepository) {
         this.personRepository = personRepository;
         this.raceRepository = raceRepository;
         this.wealthRepository = wealthRepository;
         this.abilityScoreRepository = abilityScoreRepository;
+        this.jobAssignmentRepository = jobAssignmentRepository;
+        this.eventPlacePersonRepository = eventPlacePersonRepository;
     }
 
     @Override
@@ -49,7 +53,7 @@ public class PersonService implements IPerson {
             throw new IllegalArgumentException(NULL_OR_EMPTY.message);
         }
         if (personAlreadyExists(person)) {
-            throw new DataIntegrityViolationException(NAME_EXISTS.message);
+            throw new DataIntegrityViolationException(PERSON_EXISTS.message);
         }
         if (hasForeignKeys(person) &&
                 RepositoryHelper.foreignKeyIsNotValid(personRepository, getListOfForeignKeyRepositories(), person)) {
@@ -65,7 +69,9 @@ public class PersonService implements IPerson {
         if (RepositoryHelper.cannotFindId(personRepository, personId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-        //TODO: fk check when Person is a fk
+        if (RepositoryHelper.isForeignKey(getReposWherePersonIsAForeignKey(), FK_PERSON.columnName, personId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
         personRepository.deleteById(personId);
     }
 
@@ -100,6 +106,10 @@ public class PersonService implements IPerson {
 
     private boolean personAlreadyExists(Person person) {
         return personRepository.personExists(person).isPresent();
+    }
+
+    private List<CrudRepository> getReposWherePersonIsAForeignKey() {
+        return new ArrayList<>(Arrays.asList(jobAssignmentRepository, eventPlacePersonRepository));
     }
 
     private boolean hasForeignKeys(Person person) {
